Enable SD card server in Rover and Base mode
In Base mode, the SD card server will come on shortly after the NTRIP
server starts.  The IP address of the SD card server is output to the
serial port at a regular interval, currently 15 seconds.  Using a web
browser and connecting with HTTP:// to that IP address returns a short
webpage with a link to the SD card listing page.  Clicking on a file
link will cause the file to be downloaded from the SD card to the web
browser.