Make sdCardSemaphore errors visible by switching log_d to Serial.print*
Work is getting skipped when the code fails to obtain the
sdCardSemaphore.  Switching from log_d to Serial.printf for the error
cases to make them customer visible so that the issue can be found and
fixed.  This change reviews the various cases and adds comments to each
one to indicate if they are safe or should be exposed as errors.

* menuFirmware.ino - scanForFirmware skips looking for firmware files,
this is an error when there is a firmware file on the microSD card

* menuMain.ino - factoryReset skips removing settings file from microSD
card thus failing to perform the factory reset

* menuMessages.ino - beginLogging skips opening the log file, OK since
logging remains offline and will be retried again by the next call to
updateLog in loop

* menuMessages.ino - findLastLog skips finding the last log file, this
leads to a new log file being created when data should be appended to an
existing log file

* menuSystem.ino - menuSystem skips listing the microSD card files even
though files exist on the microSD card

* NVM.ino - recordSystemSettingsToFileSD skips writing current settings
to the SD card thus the restored settings do not match what was set

* NVM.ino - loadSystemSettingsFromFileSD skips loading the current
settings thus the customer is not getting the documented behavior when a
microSD card is inserted

* RTK_Surveyor.ino - updateLogs skips closing the log file, this is OK
because on the next loop it will be retried and in the interim more data
may be written to the file and the log file will eventually get closed

* RTK_Surveyor.ino - updateLogs skips closing the log file after the log
duration is exceeded, this is OK because on the next loop it will be
retried and the log file with eventually get closed

* RTK_Surveyor.ino - updateLogs skips flushing the log file write buffer
when the periodic duration is exceeded, this is OK because on the next
loop it will be retried and the log file will eventually get flushed

* RTK_Surveyor.ino - updateLogs skips writing trigger event to log file,
while a retry occurs during the next loop, it is possible to loose
trigger events if they occur too rapidly or if the log file is closed,
changed log_d to log_w

* RTK_Surveyor.ino - updateLogs skips reporting the log file size, this
is OK because while on the next loop it will be retried

* System.ino - createTestFile fails to create the test file and thus
bring the microSD card online, this is OK because the next loop will
attempt to bring the microSD card online

* Tasks.ino - F9PSerialReadTask skips writing bytes to the log file,
this is an error where a partial binary or NEMA message could be written
to the log file.