Mount and dismount the microSD card
* Mount and dismount the microSD card in loadSystemSettingsFromFileSD.

The pre-existing error still exists:
The settings are not loaded from the microSD card when the code fails to
obtain the sdCardSemaphore thus violating the documented settings
priority!

* Mount and dismount the microSD card in recordSystemSettingsToFileSD.

The pre-existing error still exists:
When the code fails to obtain sdCardSemaphore an error occurs because
the current settings are not written to the microSD card.  Thus because
priority is given to loading settings from the microSD card the intended
settings are not properly loaded!

* Dismount the microSD card in powerDown

* Mount the microSD card in updateLogs when logging is enabled

* Dismount the microSD card in updateLogs when logging is disabled

* Mount and dismount the microSD card in menuSystem

The pre-existing error still exists:
The code fails to list the SD card files upon failure to get the
sdCardSemaphoore.

* Mount and dismount the microSD card around firmware updates
   * Added mountSDThenUpdate
   * Call mountSDThenUpdate instead of updateFromSD