Write the marks data to the marks file and display the result
The display behavior does not change when the marks file is disabled.
After the marks file is enabled, there are some new display values that
occur depending upon the state of logging and marks output.

* Both enabled & both successful: Event Marked
* Both enabled & marks failed: Not Marked
* Both enabled & not logging: No Logging
* Both enabled & both fail: Mark Failure

* Only marks enabled & successful: Marked
* Only marks enabled & failed: Not Marked

The marks file has the file name form: Marks_YYYY_MM_DD.txt

When the marks write is successful the following data is written as a
single line to the marks file with comma separated values:

* Date and time
* Longitude
* Latitude
* Altitude
* Satellites In View
* Horizontal Position Accuracy
* Battery level
* Battery voltage