Allow more than 4 profile states
Use a bit vector to specify the active profiles.  The number of profiles
may be increased to 8 without any changes.  Changing activeProfiles to a
uint32_t enables up to 32 profiles, or 64 profiles by using a uint64_t.
The uint8_t for profileNumber, displayProfile and index do not need to
be increased unless the total number of profiles exceeds 256.

The following changes were made:
* At reset, assume no profiles are active
* Change the activeProfiles variable to a bit vector
* Remove the STATE_PROFILE_* values
* Add the STATE_PROFILE value
* Change the NVM routines to support a bit vector
* Change the display routines to use the bit vector
* Change the button routine to use the bit vector
* Combine display code that displays the profiles