WiFi Config: Display button down arrows after save
There was a display bug when a section (e.g GNSS Config) was opened up
and the save button is pressed.  Prior to pressing the save button, the
arrow on the section button (e.g. GNSS Config) is point up to indicate
that the section may be collapsed.  After pressing the save button, the
section is collapsed but the arrow on the section button remains
pointing up.

This change fixes this display error by changing all of the section
buttons to have a down arrow.  Each of the caret icons is given a name
(id) and any icon direction is removed and the icon down direction is
added.