Move Bluetooth test into Debug menu
Move the Bluetooth test code into Bluetooth.ino.

Add a parameter to the test code to enable/disable running the test.

Use the common code in the System menu to just display the Bluetooth
status.

Add a menu item to the Debug menu to run the Bluetooth test.

Display Bluetooth test results all at once to allow additional messages
to be displayed during the test.