Update the NMEA client support program
Enable the NMEA client program to forward data from the RTK NMEA TCP
server to the Vespucci NMEA server.

Use this on a local WiFi network.  Both the RTK and the Android phone
need to connect to the local WiFi network.  This NMEA client program
eliminates the RTK NMEA client code and validates the Android connection
to the Vespucci NMEA server.  Once the Android connection is working, it
is easy to enable the RTK NMEA client by enabling the Mobile Hot Spot on
the phone and switching the WiFi SSID and password in the RTK.