Add information on iOS BLE to Bluetooth Page
I got stuck here for half a day trying to configure the device to work with my iPhone.  I couldn't understand why it wouldn't pair with my phone when it paired just fine with my computer.  Finally I read ahead in the instructions and at the very bottom of the GIS software page I found this nugget of information.  I almost missed it since most of the GIS software was not relevant to me.  I recommend moving this nugget of information up a page so others don't get stuck here quite so easily.