Add back custom version of BluetoothSerial
This originally comes from RTK Firmware ~v2.0. This change exposes the TX/RX buffer sizes so that we can dynamically set them (from settings) at run time. The BluetoothSerial library defaults to 512 bytes in the RX buffer. Because RTCM is ~530 to 600 bytes, there are times when the btReadTask cannot keep up and an RTCM sentence is corrupted. This change reduces the number of RTCM bytes dropped during SPP from the phone. By default, this will increase the overall RAM usage by 512 bytes.