Bug Fix: Discard the oldest data in the ring buffer
The previous behavior discarded the newest data and possibly corrupting
the data streams for the all of the devices which are keeping up with
their writes.

This fix wraps the buffer pointer for the slowest device, discarding and
entire buffer worth of data and corrupting the data stream for the
slowest device.  The data streams for the other devices are not
impacted.