Add support code to debug system hangs
Add support code to debug system hangs.  This code is disabled by
default.

To enable this code
* Isolate the failure case
* Set DEAD_MAN_WALKING_ENABLED to 1
* Place the START_DEAD_MAN_WALKING after the last good output, only
printed in the failure case

The goal of this code is to locate where in the system the hang is
occurring.  This is done by executing the START_DEAD_MAN_WALKING macro
after an isolated event that occurs just before the failure.  The macro
enables as much debug code as possible, outputs the routine names in the
loop and display states before the switch statements.  This additional
output should provide a lot of hints as to how far the code executed
after the event before the system hung.  Further debugging most likely
will be necessary, but the amount of code should be greatly reduced.